Redesign Workspace | Implementing Factory Design
