+++ home/main/res/css/workspace.css
@@ -1,0 +1,3 @@
+.scroll-pane{
+    -fx-background-color: transparent;
+}
