--- home/main/res/css/workspaceStyle.css
@@ -1,10 +1,0 @@
-.button
-{
-    -fx-background-color : #007AFF;
-    -fx-font-size: 14px;
-}
-
-
-#topBar{
-    -fx-background-color: #F3F3F3;
-}
