+++ home/main/res/view/workspace.fxml
@@ -1,96 +1,62 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
+<?import javafx.scene.Cursor?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.ScrollPane?>
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.text.Font?>
 
-<BorderPane prefHeight="940.0" prefWidth="1024.0" stylesheets="@../css/workspaceStyle.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.src.controllers.WorkspaceController">
+<BorderPane prefHeight="940.0" prefWidth="1024.0" style="-fx-background-color: #FFFFFF;" stylesheets="/main/res/css/workspace.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="main.src.controllers.WorkspaceController">
    <top>
-      <HBox prefHeight="45.0" prefWidth="200.0" style="-fx-background-color: #F3F3F3;" BorderPane.alignment="CENTER">
-            <Pane prefHeight="45.0">
-                  <Label layoutY="6.0" prefHeight="33.0" text="MathMathX">
+      <HBox prefHeight="40.0" prefWidth="200.0" style="-fx-background-color: #E3E3E3;" BorderPane.alignment="CENTER">
+         <children>
+            <Pane prefHeight="45.0" prefWidth="135.0">
+               <children>
+                  <Label fx:id="homeButton" layoutX="38.0" layoutY="7.0" text="Home">
                      <font>
-                        <Font name="Helvetica Bold" size="16.0" />
+                        <Font name="FreeSans Bold" size="14.0" />
                      </font>
                      <padding>
-                        <Insets left="10.0" right="10.0" />
+                        <Insets bottom="5.0" left="10.0" right="10.0" top="5.0" />
                      </padding>
+                     <cursor>
+                        <Cursor fx:constant="HAND" />
+                     </cursor>
                   </Label>
-               <HBox.margin>
-                  <Insets />
-               </HBox.margin>
+               </children>
             </Pane>
-            <Pane prefWidth="45.0" />
-            <Pane prefHeight="45.0">
-                  <Label layoutY="6.0" prefHeight="33.0" text="Home">
+            <Pane prefHeight="45.0" prefWidth="135.0">
+               <children>
+                  <Label layoutX="38.0" layoutY="7.0" text="Reset">
                      <font>
-                        <Font name="Helvetica" size="14.0" />
+                        <Font name="FreeSans Bold" size="14.0" />
                      </font>
                      <padding>
-                        <Insets left="10.0" right="10.0" />
+                        <Insets bottom="5.0" left="10.0" right="10.0" top="5.0" />
                      </padding>
+                     <cursor>
+                        <Cursor fx:constant="HAND" />
+                     </cursor>
                   </Label>
+               </children>
             </Pane>
-            <Pane prefWidth="45.0" />
-            <Pane prefHeight="45.0">
-                  <Label layoutY="6.0" prefHeight="33.0" text="Reset">
-                     <font>
-                        <Font name="Helvetica" size="14.0" />
-                     </font>
-                     <padding>
-                        <Insets left="10.0" right="10.0" />
-                     </padding>
-                  </Label>
-            </Pane>
+         </children>
       </HBox>
    </top>
-   <center>
+   <left>
       <Pane prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER">
-
-            <StackPane fx:id="operator" layoutX="232.0" layoutY="220.0">
-
-                  <Rectangle fx:id="rectangle" arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="70.0" opacity="0.8" stroke="BLACK" strokeLineCap="ROUND" strokeLineJoin="ROUND" strokeType="INSIDE" strokeWidth="2.0" style="-fx-arc-width: 70; -fx-arc-height: 70;" width="190" />
-                  <HBox alignment="CENTER" minWidth="190.0">
-                        <TextField fx:id="number1" minWidth="40.0" prefWidth="40.0">
-                           <HBox.margin>
-                              <Insets bottom="10.0" left="20.0" right="10.0" top="10.0" />
-                           </HBox.margin>
-                        </TextField>
-                        <Label alignment="CENTER" contentDisplay="TOP" maxHeight="40.0" maxWidth="30.0" minHeight="30.0" minWidth="30.0" prefHeight="30.0" prefWidth="30.0" text="+" textAlignment="CENTER" textFill="WHITE">
-                           <font>
-                              <Font size="30.0" />
-                           </font>
-                        </Label>
-                        <TextField fx:id="number2" minWidth="40.0" prefWidth="40.0">
-                           <HBox.margin>
-                              <Insets bottom="10.0" left="10.0" right="20.0" top="10.0" />
-                           </HBox.margin>
-                        </TextField>
-                     
-                  </HBox>
-               
-            </StackPane>
-         
-      </Pane>
+         <children>
+            <ScrollPane prefHeight="900.0" prefWidth="200.0">
+               <content>
+                  <VBox fx:id="expressionPane" prefHeight="200.0" prefWidth="200.0" />
+               </content></ScrollPane>
+         </children></Pane>
+   </left>
+   <center>
+      <Pane prefHeight="200.0" prefWidth="200.0" style="-fx-background-color: #FFFFFF" BorderPane.alignment="CENTER" />
    </center>
-   <bottom>
-      <VBox prefHeight="60.0" prefWidth="1024.0" BorderPane.alignment="CENTER">
-         
-            <Label prefHeight="60.0" prefWidth="1024.0" text="Result">
-               <font>
-                  <Font name="System Bold" size="19.0" />
-               </font></Label>
-         
-         <padding>
-            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-         </padding>
-      </VBox>
-   </bottom>
 </BorderPane>
