+++ home/main/src/controllers/DoubleInputExpression.java
@@ -1,0 +1,13 @@
+package main.src.controllers;
+
+import javafx.scene.control.Button;
+import javafx.scene.layout.VBox;
+
+public class DoubleInputExpression implements Expression{
+
+    @Override
+    public void produceExpression(VBox vBox) {
+        Button button=new Button("Hello");
+
+    }
+}
