+++ home/main/src/controllers/Expression.java
@@ -1,0 +1,7 @@
+package main.src.controllers;
+
+import javafx.scene.layout.VBox;
+
+public interface Expression {
+    void produceExpression(VBox vBox);
+}
