+++ home/main/src/controllers/ExpressionFactory.java
@@ -1,0 +1,21 @@
+package main.src.controllers;
+
+import javafx.scene.layout.VBox;
+
+public abstract class ExpressionFactory {
+
+    public static Expression produceExpression(WorkspaceController.expressionType expressionType,VBox vBox){
+        Expression expression=null;
+        switch (expressionType){
+            case SINGLE: {
+                expression=new SingleInputExpression();
+                break;
+            }
+            case DOUBLE:{
+                expression=new SingleInputExpression();
+                break;
+            }
+        }
+        return expression;
+    }
+}
